+++ src/ExtensionsHost/CExtensionsHost.cpp
@@ -5,6 +5,20 @@
 
 #include "stdafx.h"
 
+#include "../ExtensionsCommon/ModuleUtils.h"
+#include "../ExtensionsCommon/PathUtils.h"
+#include "../ExtensionsCommon/TextUtils.h"
+
+#include "../ExtensionsCommon/AgentValidation.h"
+#include "../ExtensionsCommon/Environment.h"
+
+#include "../ExtensionsHostLib/RawProfilerHookLoader.h"
+#include "../ExtensionsHostLib/RawProfilerHookSettingsReader.h"
+#include "../ExtensionsHostLib/SafeFindFileHandle.h"
+
+#include "../InstrumentationEngine/ImplQueryInterface.h"
+#include "../InstrumentationEngine/refcount.h"
+
 #include "CExtensionsHost.h"
 
 using Agent::Diagnostics::Param;