+++ tests/InstrEngineTests/InstrEngineTests/TestParameters.cs
@@ -1,5 +1,5 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// 
+//
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
@@ -12,8 +12,7 @@ namespace InstrEngineTests
 
         public static void Initialize(TestContext context)
         {
-            // The NaglerProfilerHost disregards code signing verification.
-            // However, InstrumentationEngine now has a built-in ExtensionsHost
+            // InstrumentationEngine uses a built-in default ExtensionsHost
             // which requires code sign verification. For tests, we disable verification.
             DisableMethodSignatureValidation = true;
         }