+++ tests/InstrEngineTests/TestInstrumentationMethod/InstrumentationMethod.h
@@ -1,5 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
-// 
+//
 
 #pragma once
 
@@ -68,7 +68,7 @@ public:
         COM_INTERFACE_ENTRY(IInstrumentationMethodExceptionEvents)
     END_COM_MAP()
 
-    DECLARE_REGISTRY_RESOURCEID(IDR_CPROFILERHOST)
+    DECLARE_REGISTRY_RESOURCEID(IDR_CINSTRMETHOD)
 
     // Ensure only a single instance of CInstrumentationMethod is created
     DECLARE_CLASSFACTORY_SINGLETON(CInstrumentationMethod)